Refactor `Parser` into commons
