+++ src/main/java/spleetwaise/address/logic/parser/DeleteCommandParser.java
@@ -4,7 +4,8 @@ import static spleetwaise.address.logic.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
 
 import spleetwaise.address.commons.core.index.Index;
 import spleetwaise.address.logic.commands.DeleteCommand;
-import spleetwaise.address.logic.parser.exceptions.ParseException;
+import spleetwaise.commons.logic.parser.Parser;
+import spleetwaise.commons.logic.parser.exceptions.ParseException;
 
 /**
  * Parses input arguments and creates a new DeleteCommand object
