+++ src/main/java/spleetwaise/address/logic/parser/ParserUtil.java
@@ -8,13 +8,13 @@ import java.util.Set;
 
 import spleetwaise.address.commons.core.index.Index;
 import spleetwaise.address.commons.util.StringUtil;
-import spleetwaise.address.logic.parser.exceptions.ParseException;
 import spleetwaise.address.model.person.Address;
 import spleetwaise.address.model.person.Email;
 import spleetwaise.address.model.person.Name;
 import spleetwaise.address.model.person.Phone;
 import spleetwaise.address.model.person.Remark;
 import spleetwaise.address.model.tag.Tag;
+import spleetwaise.commons.logic.parser.exceptions.ParseException;
 
 /**
  * Contains utility methods used for parsing strings in the various *Parser classes.
