+++ src/main/java/spleetwaise/commons/logic/parser/Parser.java
@@ -1,7 +1,7 @@
-package spleetwaise.address.logic.parser;
+package spleetwaise.commons.logic.parser;
 
-import spleetwaise.address.logic.parser.exceptions.ParseException;
 import spleetwaise.commons.logic.commands.Command;
+import spleetwaise.commons.logic.parser.exceptions.ParseException;
 
 /**
  * Represents a Parser that is able to parse user input into a {@code Command} of type {@code T}.
@@ -11,7 +11,7 @@ public interface Parser<T extends Command> {
     /**
      * Parses {@code userInput} into a command and returns it.
      *
-     * @throws ParseException if {@code userInput} does not conform the expected format
+     * @throws ParseException if {@code userInput} does not conform the expected format.
      */
     T parse(String userInput) throws ParseException;
 }
