+++ src/test/java/spleetwaise/address/logic/parser/CommandParserTestUtil.java
@@ -2,8 +2,9 @@ package spleetwaise.address.logic.parser;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-import spleetwaise.address.logic.parser.exceptions.ParseException;
 import spleetwaise.commons.logic.commands.Command;
+import spleetwaise.commons.logic.parser.Parser;
+import spleetwaise.commons.logic.parser.exceptions.ParseException;
 
 /**
  * Contains helper methods for testing command parsers.
